fix: Multiple UI and display issues
- Fix notes displaying raw HTML tags by rendering editor content properly
- Fix "Okonomi" typo to "Økonomi" in dashboard shortcuts
- Fix decimal days display (e.g. "116.494..." to "116") by casting
  diffInDays() results to int in multiple models and components
- Fix VoucherManager dependency injection error by adding default
  parameter values to openModal method

+++ app/Livewire/AltinnDashboard.php
@@ -175,7 +175,7 @@ class AltinnDashboard extends Component
 
         foreach ($deadlines as $deadline) {
             if ($deadline['status']['code'] !== 'accepted' && $deadline['deadline']->isFuture()) {
-                $daysUntil = now()->diffInDays($deadline['deadline'], false);
+                $daysUntil = (int) now()->diffInDays($deadline['deadline'], false);
                 $deadline['days_until'] = $daysUntil;
                 $deadline['urgency'] = $this->getUrgencyLevel($daysUntil);
                 $upcoming[] = $deadline;
@@ -195,7 +195,7 @@ class AltinnDashboard extends Component
 
         foreach ($deadlines as $deadline) {
             if ($deadline['status']['code'] !== 'accepted' && $deadline['deadline']->isPast()) {
-                $daysPast = now()->diffInDays($deadline['deadline']);
+                $daysPast = (int) now()->diffInDays($deadline['deadline']);
                 $deadline['days_past'] = $daysPast;
                 $overdue[] = $deadline;
             }
