+++ app/Models/AnnualAccount.php
@@ -197,7 +197,7 @@ class AnnualAccount extends Model
 
     public function getDaysUntilDeadline(): int
     {
-        return now()->diffInDays($this->getDeadline(), false);
+        return (int) now()->diffInDays($this->getDeadline(), false);
     }
 
     public function isOverdue(): bool
