+++ app/Models/ShareholderReport.php
@@ -128,7 +128,7 @@ class ShareholderReport extends Model
 
     public function getDaysUntilDeadline(): int
     {
-        return now()->diffInDays($this->getDeadline(), false);
+        return (int) now()->diffInDays($this->getDeadline(), false);
     }
 
     public function isOverdue(): bool
